Create README.md
A online Bus Reservation System Project In Java will calculate the tax amount and road amount and total passenger number and gives you the detail about the ticket fare. Also, you can select the type of ticket you want to get like the First Class or Economy class. Also, you can select the ticket mode that is whether you want to return with the same bus or not.